pyproject.toml
[project]
name = "genesis-robotics"
version = "0.1.0"
description = ""
authors = [
    {name = "OtavioRMC",email = "otavioribeiromoreiradacosta@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "torch (>=2.8.0,<3.0.0)",
    "pymeshlab (==2023.12.post3)",
    "genesis-world (>=0.3.0,<0.4.0)"
  
]

[tool.poetry]
packages = [{include = "genesis_robotics", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
